Fix gaps in road
White lines would occasionally flash on the road. This was caused by
texture bleeding: the crossroad tile bled into the main road tile.

To avoid this, add a 2px frame around tiles filed with the same
pixels as the borders of the tiles. This frame is generated with
tools/pad-map-tiles.