Fix reward unlocked screen sometimes appearing when it should not
Sometimes the reward unlocked screen appeared for already unlocked rewards.

It seems this only happened after a refresh (using F5 to debug a screen)
because it caused the recreation of all rewards, which made the set of
already unlocked rewards kept in Maestro not match any of the new rewards.
Keep this state in RewardManager to fix this.