Fix crash when starting a race if PW_ASSETS_DIR is set
Tthe TmxMapLoader assumes dependent files are siblings of the tmx file.

This causes a crash if PW_ASSETS_DIR is set when starting a race from a
checkout if game assets have not been processed, because in this case
the texture files are not in the same directory as the tmx file.