Make sure rocket car cannot get stuck on thin walls
When hitting a thin wall at the correct angle, it was possible for a front
wheel to go on one side of the wall and the body on the other side.

Add a shape to the body to "protect" the front wheels so that they can't be
separated from the body.