Fix vehicle falling on the other side of the bridge after drop
Before this change, a vehicle being dropped would not respond to collisions
until it had fully landed. This could cause it to go through the wall of
the bridge on the "River" track, as reported in bug 302.

Reactivate collisions as soon as the vehicle is dropped to avoid this.

Fixes #302