Refactor NLog: now Gdx logs go through NLog instead of NLog using Gdx logs
The reason for this change is that Gdx logs can appear in the wrong order
on the desktop. This is because info and debug level logs are sent to
stdout, but warning and error logs are sent to stderr. This creates
confusion and makes log-based debugging harder, if not impossible.