Fix incorrect player rank on 2nd or 3rd race of championship
The loop to get the player rank by comparing it to other racers stopped
after it found the player racer. This is wrong: racers after the player in
 mRacers may be in front of them in the race. This is because mRacers is no
longer sorted in every frame.

This bug is not visible on the 1st race of a championship or in quick race
 because on the first race the player is at the end of mRacers, so
getRacerRank() compares it with all other racers.